requirement.txt
torch>=1.9.0
torch-geometric>=2.0.0
scikit-learn>=0.24.0
wandb>=0.12.0
tqdm>=4.60.0
numpy>=1.19.0
PyYAML>=5.4.0 # For optional YAML config loading
scipy>=1.6.0 # For FastGCN sampler (sparse matrix operations)